refactor(server): move parseQueueListFromEnv to a dedicated file
Address review feedback: extract parseQueueListFromEnv into its own util
file (with its own spec) instead of colocating it with
shouldCreateWorkerForQueue.
