+++ packages/twenty-server/src/engine/core-modules/message-queue/utils/__tests__/should-create-worker-for-queue.util.spec.ts
@@ -1,21 +1,5 @@
 import { MessageQueue } from 'src/engine/core-modules/message-queue/message-queue.constants';
-import {
-  parseQueueListFromEnv,
-  shouldCreateWorkerForQueue,
-} from 'src/engine/core-modules/message-queue/utils/should-create-worker-for-queue.util';
-
-describe('parseQueueListFromEnv', () => {
-  it('should return an empty array when the value is undefined', () => {
-    expect(parseQueueListFromEnv(undefined)).toEqual([]);
-  });
-
-  it('should split, trim and drop empty entries', () => {
-    expect(parseQueueListFromEnv('workspace-queue, workflow-queue ,')).toEqual([
-      'workspace-queue',
-      'workflow-queue',
-    ]);
-  });
-});
+import { shouldCreateWorkerForQueue } from 'src/engine/core-modules/message-queue/utils/should-create-worker-for-queue.util';
 
 describe('shouldCreateWorkerForQueue', () => {
   it('should create a worker for every queue when no filters are set', () => {
