+++ packages/twenty-server/src/engine/core-modules/message-queue/utils/parse-queue-list-from-env.util.ts
@@ -1,0 +1,5 @@
+export const parseQueueListFromEnv = (value: string | undefined): string[] =>
+  value
+    ?.split(',')
+    .map((queueName) => queueName.trim())
+    .filter((queueName) => queueName.length > 0) ?? [];
