+++ packages/twenty-server/src/engine/core-modules/message-queue/utils/should-create-worker-for-queue.util.ts
@@ -1,11 +1,5 @@
 import { type MessageQueue } from 'src/engine/core-modules/message-queue/message-queue.constants';
 
-export const parseQueueListFromEnv = (value: string | undefined): string[] =>
-  value
-    ?.split(',')
-    .map((queueName) => queueName.trim())
-    .filter((queueName) => queueName.length > 0) ?? [];
-
 export const shouldCreateWorkerForQueue = ({
   queueName,
   enabledQueues,
